ARW: clone() a much smaller amount of data for WB
Since sony "decryption" requires parsing a TIFF after "decrypting" it
and since TIFF's have absolute offsets the code resorted to duplicating
the whole buffer, modifying it and parsing the TIFF. Instead have TIFF
parsing take a base_offset parameters that gets subtracted from the
offsets so we can parse from a partial buffer only. This works as long
as the Tiff doesn't have links to outside the data range itself which
thankfully seems to be the case for ARW WB metadata.