Speedup the bitpump by reading 32bits at a time
The bitpump was simplified just like the rawspeed one in that it always
fetched from the buffer no matter how few bits it needed. This is probably
a good idea in a language without bounds checks but in rust reading 32
bits at a time and consuming those makes for a nice speedup.