Set number of decode threads by number of cores
On linux use /proc/cpuinfo to determine the number of physical cores.
If that fails use the num_cpu crate which will return the number of
logical cpus. A PR has been sent to add this feature to num_cpu and
there are already implementations for other OSs as well.