imageops: make demosaic the first operation
Since demosaic is usually more expensive than most operations do it
first so settings changes in the other operations don't require
redoing the demosaic. While we're at it make levels and whitebalance
a single operation so we're not doing multiple passes of the array.
Performance seems to stay about the same.