imageops: get rid of ingest and scale demosaic
Instead of converting to f32 as a separate op (and spending 150ms
on it for a 24MP image) just do the f32 conversion in demosaic and
the [0;1] conversion in level_and_balance. While we're at it add
a scaled demosaic step that allows us to do 378x251 thumbnails from
a 24MP image in 200ms.