Replace homebrew parallelization with rayon
Instead of doing my own mix of crossbeam+num_cpus that requires doing
an explicit tuning for the number of CPUs, use rayon and let it handle
that automatically. The benchmarking is equivalent to the previous
solution if using all the CPUs and not just the physical ones but now
this should be more reasonable as rayon's work-stealing setup should
make this friendlier when the program is already doing other stuff.