Use my fixed afl.rs to fully fix panic handling
Instead of removing the hook on every iteration use a new fuzz_nohook!
to not install it in the first place. This should be a performance boost
but more importantly makes stability 100% which should help with fuzzing
efficiency both during the run and by making cmin more effective at
pruning out unneeded files.