Implement efficient fuzzing strategy for decoders
Fuzzing whole formats is tricky because it's not possible to create a
raw file that's exactly the same as what the camera would create but
just 32x32 or something like that. So most of the fuzzing effort when
using a complete file is in the metadata parsing. But the decoders
themselves are particularly important to fuzz completely to make sure
there are no infinite loops anywhere.

To exercise the decoders themselves more completely implement a fuzzing
target that exercises them directly. This is done by exposing the
decoders without any parsing around them in a very simple raw format:

- starts with three 16bit integers identifying format/width/height
- continues with any extra metadata when the format requires it
- finally the actual raw data comes at the end

This then allows limiting the image to 64x64 at most and with that
getting faster fuzzing and more coverage of the actual decoders.