Implement decoding for split encoded NEF
Some NEF files have a split encoding where the second part of the image
uses a special mode where the differences are shifted. Adapt the Huffman
decoder to support this specifically for NEF. Huffman BigTable is lost
for now for NEF. It should probably be reimplemented with some better
code anyway.

Fixes #34