Generalize the small table code bit length
When the big table doesn't work there's a table of huffman lengths to
use before going down the slow path. Make that code generic in the
length of the code used so be able to bump it up if needed. Preliminary
testing doesn't show any big advantage to going larger but the picture
may change when other tuning is done.