CR3: Add CRM support by utilizing STSZ sample count
CRM movie files uses the same CR3/BMFF structure.

For each recorded frame, the frame is encoded by Crx codec and
written as sample. So a CRM file with 2 seconds duration and 30 fps
contains 60 samples inside trak 1. Each sample can independently
be decoded.

The order of trak boxes is different. The main (video frame) trak is
the first one. We can use CCTP box to find correct trak id.
The trak for CTMD is not listed in CCTP but it exists?!